--- a/binary/jce/structs.go
+++ b/binary/jce/structs.go
@@ -555,10 +555,12 @@ func (pkt *RequestPacket) ReadFrom(r *JceReader) {
 	pkt.IRequestId = r.ReadInt32(4)
 	pkt.SServantName = r.ReadString(5)
 	pkt.SFuncName = r.ReadString(6)
-	r.ReadSlice(&pkt.SBuffer, 7)
+	pkt.SBuffer = r.ReadBytes(7)
 	pkt.ITimeout = r.ReadInt32(8)
-	r.ReadMap(pkt.Context, 9)
-	r.ReadMap(pkt.Status, 10)
+	// r.ReadMap(pkt.Context, 9)
+	pkt.Context = r.ReadMapStrStr(9)
+	// r.ReadMap(pkt.Status, 10)
+	pkt.Status = r.ReadMapStrStr(10)
 }
 
 func (pkt *RequestDataVersion3) ToBytes() []byte {
@@ -568,8 +570,7 @@ func (pkt *RequestDataVersion3) ToBytes() []byte {
 }
 
 func (pkt *RequestDataVersion3) ReadFrom(r *JceReader) {
-	pkt.Map = make(map[string][]byte)
-	r.ReadMap(pkt.Map, 0)
+	pkt.Map = r.ReadMapStrByte(0)
 }
 
 func (pkt *RequestDataVersion2) ToBytes() []byte {
@@ -579,8 +580,7 @@ func (pkt *RequestDataVersion2) ToBytes() []byte {
 }
 
 func (pkt *RequestDataVersion2) ReadFrom(r *JceReader) {
-	pkt.Map = make(map[string]map[string][]byte)
-	r.ReadMap(pkt.Map, 0)
+	pkt.Map = r.ReadMapStrMapStrByte(0)
 }
 
 func (pkt *SsoServerInfo) ReadFrom(r *JceReader) {
@@ -590,22 +590,15 @@ func (pkt *SsoServerInfo) ReadFrom(r *JceReader) {
 }
 
 func (pkt *FileStoragePushFSSvcList) ReadFrom(r *JceReader) {
-	pkt.UploadList = []FileStorageServerInfo{}
-	pkt.PicDownloadList = []FileStorageServerInfo{}
-	pkt.GPicDownloadList = []FileStorageServerInfo{}
-	pkt.QZoneProxyServiceList = []FileStorageServerInfo{}
-	pkt.UrlEncodeServiceList = []FileStorageServerInfo{}
+	pkt.UploadList = r.ReadFileStorageServerInfos(0)
+	pkt.PicDownloadList = r.ReadFileStorageServerInfos(1)
+	pkt.GPicDownloadList = r.ReadFileStorageServerInfos(2)
+	pkt.QZoneProxyServiceList = r.ReadFileStorageServerInfos(3)
+	pkt.UrlEncodeServiceList = r.ReadFileStorageServerInfos(4)
 	pkt.BigDataChannel = &BigDataChannel{}
-	pkt.VipEmotionList = []FileStorageServerInfo{}
-	pkt.C2CPicDownList = []FileStorageServerInfo{}
-	r.ReadSlice(&pkt.UploadList, 0)
-	r.ReadSlice(&pkt.PicDownloadList, 1)
-	r.ReadSlice(&pkt.GPicDownloadList, 2)
-	r.ReadSlice(&pkt.QZoneProxyServiceList, 3)
-	r.ReadSlice(&pkt.UrlEncodeServiceList, 4)
+	pkt.VipEmotionList = r.ReadFileStorageServerInfos(5)
+	pkt.C2CPicDownList = r.ReadFileStorageServerInfos(7)
 	r.ReadJceStruct(pkt.BigDataChannel, 5)
-	r.ReadSlice(&pkt.VipEmotionList, 6)
-	r.ReadSlice(&pkt.C2CPicDownList, 7)
 	pkt.PttList = r.ReadBytes(10)
 }
 
@@ -615,8 +608,7 @@ func (pkt *FileStorageServerInfo) ReadFrom(r *JceReader) {
 }
 
 func (pkt *BigDataChannel) ReadFrom(r *JceReader) {
-	pkt.IPLists = []BigDataIPList{}
-	r.ReadSlice(&pkt.IPLists, 0)
+	pkt.IPLists = r.ReadBigDataIPLists(0)
 	pkt.SigSession = r.ReadBytes(1)
 	pkt.KeySession = r.ReadBytes(2)
 	pkt.SigUin = r.ReadInt64(3)
@@ -625,9 +617,8 @@ func (pkt *BigDataChannel) ReadFrom(r *JceReader) {
 }
 
 func (pkt *BigDataIPList) ReadFrom(r *JceReader) {
-	pkt.IPList = []BigDataIPInfo{}
 	pkt.ServiceType = r.ReadInt64(0)
-	r.ReadSlice(&pkt.IPList, 1)
+	pkt.IPList = r.ReadBigDataIPInfos(1)
 	pkt.FragmentSize = r.ReadInt64(3)
 }
 
@@ -692,8 +683,7 @@ func (pkt *FriendInfo) ReadFrom(r *JceReader) {
 	pkt.Nick = r.ReadString(14)
 	pkt.Network = r.ReadByte(20)
 	pkt.NetworkType = r.ReadInt32(24)
-	pkt.CardID = []byte{}
-	r.ReadObject(&pkt.CardID, 41)
+	pkt.CardID = r.ReadBytes(41)
 }
 
 func (pkt *TroopListRequest) ToBytes() []byte {
@@ -750,8 +740,7 @@ func (pkt *PushMessageInfo) ReadFrom(r *JceReader) {
 
 func (pkt *SvcDevLoginInfo) ReadFrom(r *JceReader) {
 	pkt.AppId = r.ReadInt64(0)
-	pkt.Guid = []byte{}
-	r.ReadSlice(&pkt.Guid, 1)
+	pkt.Guid = r.ReadBytes(1)
 	pkt.LoginTime = r.ReadInt64(2)
 	pkt.LoginPlatform = r.ReadInt64(3)
 	pkt.LoginLocation = r.ReadString(4)
@@ -763,7 +752,6 @@ func (pkt *SvcDevLoginInfo) ReadFrom(r *JceReader) {
 }
 
 func (pkt *SvcRespParam) ReadFrom(r *JceReader) {
-	pkt.OnlineInfos = []OnlineInfo{}
 	pkt.PCStat = r.ReadInt32(0)
 	pkt.IsSupportC2CRoamMsg = r.ReadInt32(1)
 	pkt.IsSupportDataLine = r.ReadInt32(2)
@@ -771,7 +759,7 @@ func (pkt *SvcRespParam) ReadFrom(r *JceReader) {
 	pkt.IsSupportViewPCFile = r.ReadInt32(4)
 	pkt.PcVersion = r.ReadInt32(5)
 	pkt.RoamFlag = r.ReadInt64(6)
-	r.ReadSlice(&pkt.OnlineInfos, 7)
+	pkt.OnlineInfos = r.ReadOnlineInfos(7)
 	pkt.PCClientType = r.ReadInt32(8)
 }
 
@@ -797,7 +785,6 @@ func (pkt *OnlineInfo) ReadFrom(r *JceReader) {
 }
 
 func (pkt *SvcReqMSFLoginNotify) ReadFrom(r *JceReader) {
-	pkt.InstanceList = []InstanceInfo{}
 	pkt.AppId = r.ReadInt64(0)
 	pkt.Status = r.ReadByte(1)
 	pkt.Tablet = r.ReadByte(2)
@@ -806,7 +793,7 @@ func (pkt *SvcReqMSFLoginNotify) ReadFrom(r *JceReader) {
 	pkt.Info = r.ReadString(5)
 	pkt.ProductType = r.ReadInt64(6)
 	pkt.ClientType = r.ReadInt64(7)
-	r.ReadSlice(&pkt.InstanceList, 8)
+	pkt.InstanceList = r.ReadInstanceInfos(8)
 }
 
 func (pkt *InstanceInfo) ReadFrom(r *JceReader) {